Deleted corva/utils.py
