+++ src/corva/models/base.py
@@ -5,10 +5,9 @@ from typing import Any, Generic, List, Optional, TypeVar, Union
 
 import pydantic.generics
 
-from corva import utils
 from corva.api import Api
 from corva.configuration import Settings
-from corva.state.redis_state import RedisState
+from corva.state.redis_state import RedisState, get_cache
 
 
 class EventConfig:
@@ -55,7 +54,7 @@ class BaseContext(pydantic.generics.GenericModel, Generic[RawBaseEventTV]):
         if self.cache_:
             return self.cache_
 
-        self.cache_ = utils.get_cache(
+        self.cache_ = get_cache(
             asset_id=self.event.asset_id,
             app_stream_id=self.event.app_stream_id,
             app_connection_id=self.event.app_connection_id,
